requirements.txt
Flask==3.0.3
pymongo==4.4.0
python-dotenv==1.0.1
Werkzeug==3.0.2
